requirements.txt
langchain-community 
langchain 
crewai 
langchain-openai 
duckduckgo-search
langchain-core
langchain-google-vertexai
langchain_google_genai
streamlit